fix: secure JWT validation and storage hardening
- Integrated \`jsonwebtoken\` for signature verification in MCP server.
- Added \`MCP_OAUTH_TOKEN_SECRET\` for secure HMAC token validation.
- Hardened Turso storage with parameterized SQL (LIMIT, json_extract, etc.).
- Refactored storage modules to comply with 500-line limit per file.
- Introduced \`MAX_TAGS_PER_QUERY\` and result clamping for DoS prevention.
- Switched to async filesystem operations in storage initialization.
- Added comprehensive OAuth security integration tests.
- Updated documentation with security best practices.
