+++ agent_docs/LESSONS.md
@@ -25,9 +25,3 @@ Compact log for non-obvious workflow learnings. Pair each entry here with a shor
 - Issue: The repo's best near-term CPU/token wins were easy to miss because they are split across plans and code placeholders.
 - Root Cause: Performance opportunities span both implementation gaps and prompt-assembly workflow docs.
 - Solution: Prioritize cached retrieval wiring, QueryCache contention, bounded context assembly, hierarchical reranking, and compact high-frequency skill/docs before larger research features.
-
-## LESSON-005: Secure JWT validation requires signature verification
-
-- Issue: The MCP server used simplified JWT parsing that skipped signature verification, allowing token forgery.
-- Root Cause: Development-only validation logic was left in place without enforcing secure production defaults.
-- Solution: Use robust libraries like `jsonwebtoken` for verification. Enforce mandatory secret configuration for production modes and include security tests for signature and claim (iss, aud, exp, sub) validation.
