+++ memory-mcp/Cargo.toml
@@ -57,7 +57,6 @@ embeddings = []
 streaming-impl = []
 
 [dev-dependencies]
-jsonwebtoken = "9.3.0"
 tempfile = { workspace = true }
 tokio = { workspace = true, features = ["test-util"] }
 tracing-subscriber = { workspace = true }
